Crate ipc

Inter-process communication library for Rust

This library houses some inter-process communication utilities available across many platforms. The current primary interface is a Semaphore which is modeled after the std::sync::Semaphore structure.

Warning: This crate is only compatible with libnative currently. Crates taking advantage of libgreen will not be able to use this crate. To work around this, a helper thread may be necessary.

Example

use ipc::Semaphore;

let s = match Semaphore::new("my-fun-semaphore", 1) {
    Ok(sem) => sem,
    Err(s) => panic!("failed to create a semaphore: {}", s)
};

// lock the semaphore
let guard = s.access();

// unlock the semaphore
drop(guard);

// manage the semaphore count manually
s.acquire();
s.release();

Structs

Guard

An RAII guard used to release a semaphore automatically when it falls out of scope.

Semaphore

An atomic counter which can be shared across processes.