ipc
Struct Semaphore

pub struct Semaphore { /* private fields */ }

An atomic counter which can be shared across processes.

This counter will block the current process in access or acquire when the count is 0, waiting for a process to invoke release through some mechanism.

Implementations

impl Semaphore

pub fn new(name: &str, cnt: usize) -> Result<Semaphore>

Creates a new semaphore with the given name and count.

If the current system has no semaphore named name, then a new semaphore will be created with the initial count cnt.

If the current system already has a semaphore named name, then a handle to that semaphore will be returned and cnt will be ignored.

Note that the name provided will be mangled as necessary when passed to the underlying system, so the name is not necessarily compatible with other processes using semaphores.

Errors

Any errors which occur when creating a semaphore are returned in string form.

Example

use ipc::Semaphore;

// sem1/sem2 are handles to the same semaphore
let sem1 = Semaphore::new("foo", 1).unwrap();
let sem2 = Semaphore::new("foo", 1 /* ignored */).unwrap();

pub fn acquire(&self)

Acquire a resource of this semaphore.

This function will block until a resource is available (a count > 0), and then decrement it and return.

pub fn try_acquire(&self) -> bool

Attempt to acquire a resource of this semaphore.

This function is identical to acquire except that it will never blocked. This function returns true if a resource was acquired or false if one could not be acquired.

pub fn release(&self)

Release a resource of this semaphore.

This function will increment the count of this semaphore, waking up any waiters who would like the resource.

pub fn access(&self) -> Guard<'_>

Access a resource of this semaphore in a constrained scope.

This function will first acquire a resource and then return an RAII guard structure which will release the resource when it falls out of scope. For a mutex-like semaphore, it is recommended to use this method rather than the acquire or release methods.

pub fn try_access(&self) -> Option<Guard<'_>>

Attempt to access a resource of this semaphore.

This function is identical to access except that it will never block.